+++ substratevm/CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog summarizes major changes to GraalVM Native Image.
 * (GR-37606) Add support for URLs and short descriptions to `Feature`. This info is shown as part of the build output.
 * (GR-38965) Heap dumps are now supported in Community Edition. 
 * (GR-38951) Add `-XX:+DumpHeapAndExit` option to dump the initial heap of a native executable.
+* (GR-37582) Run image-builder on module-path per default. Opt-out with env setting `USE_NATIVE_IMAGE_JAVA_PLATFORM_MODULE_SYSTEM=false`.
 
 ## Version 22.1.0
 * (GR-36568) Add "Quick build" mode, enabled through option `-Ob`, for quicker native image builds.